// contentTypes.d.ts
export type GreenButtonContentType = 'ApplicationInformation' | 'Authorization' | 'BatchList' | 'Customer' | 'CustomerAccount' | 'CustomerAgreement' | 'ElectricPowerUsageSummary' | 'EnergyUsageSummary' | 'IntervalBlock' | 'LocalTimeParameters' | 'Meter' | 'MeterReading' | 'ReadingType' | 'ServiceLocation' | 'ServiceStatus' | 'ServiceSupplier' | 'UsagePoint' | 'UsageSummary';
export interface GreenButtonJson {
    title: string;
    updatedDate: Date;
    link: string;
    entries: GreenButtonEntry[];
}
export interface GreenButtonEntry {
    id: string;
    title: string;
    link: string;
    publishedDate: Date;
    content: GreenButtonContent;
}
export interface BaseContent {
    contentType: GreenButtonContentType;
}
export interface ApplicationInformationContent extends BaseContent {
    contentType: 'ApplicationInformation';
    dataCustodianApplicationStatus: keyof typeof lookups.dataCustodianApplicationStatuses;
    dataCustodianApplicationStatusValue?: (typeof lookups.dataCustodianApplicationStatuses)[keyof typeof lookups.dataCustodianApplicationStatuses];
    dataCustodianResourceEndpoint: objectTypes.urlString;
    authorizationServerAuthorizationEndpoint: objectTypes.urlString;
    authorizationServerTokenEndpoint: objectTypes.urlString;
    redirect_uri: objectTypes.urlString;
    thirdPartyNotifyUri: objectTypes.urlString;
    client_id: string;
    client_secret: string;
    registration_access_token: string;
    thirdPartyApplicationDescription?: string;
    thirdPartyApplicationStatus?: keyof typeof lookups.thirdPartyApplicationStatuses;
    thirdPartyApplicationStatus_value?: (typeof lookups.thirdPartyApplicationStatuses)[keyof typeof lookups.thirdPartyApplicationStatuses];
    thirdPartyApplicationType?: keyof typeof lookups.thirdPartyApplicationTypes;
    thirdPartyApplicationType_value?: (typeof lookups.thirdPartyApplicationTypes)[keyof typeof lookups.thirdPartyApplicationTypes];
    thirdPartyApplicationUse?: keyof typeof lookups.thirdPartyApplicationUses;
    thirdPartyApplicationUse_value?: (typeof lookups.thirdPartyApplicationUses)[keyof typeof lookups.thirdPartyApplicationUses];
    thirdPartyPhone?: string;
    authorizationServerUri?: objectTypes.urlString;
    authorizationServerRegistrationEndpoint?: objectTypes.urlString;
    dataCustodianBulkRequestURI: objectTypes.urlString;
    thirdPartyScopeSelectionScreenURI: objectTypes.urlString;
    thirdPartyUserPortalScreenURI?: objectTypes.urlString;
    logo_uri?: objectTypes.urlString;
    client_name: string;
    client_uri?: objectTypes.urlString;
    tos_uri?: objectTypes.urlString;
    policy_uri?: objectTypes.urlString;
    software_id: string;
    software_version: string;
    client_id_issued_at: objectTypes.timestampNumber;
    client_secret_expires_at: objectTypes.timestampNumber;
    contacts?: string[];
    token_endpoint_auth_method: string;
    scope: string[];
    grant_types: Array<keyof typeof lookups.grantTypes>;
    grant_types_values?: Array<(typeof lookups.grantTypes)[keyof typeof lookups.grantTypes]>;
    response_types: keyof typeof lookups.responseTypes;
    response_types_value?: (typeof lookups.responseTypes)[keyof typeof lookups.responseTypes];
    registration_client_uri: objectTypes.urlString;
    dataCustodianId: string;
    dataCustodianScopeSelectionScreenURI: objectTypes.urlString;
}
export interface AuthorizationContent extends BaseContent {
    contentType: 'Authorization';
    authorizedPeriod?: objectTypes.GreenButtonDuration;
    publishedPeriod?: objectTypes.GreenButtonDuration;
    status: keyof typeof lookups.authorizationStatuses;
    status_value?: (typeof lookups.authorizationStatuses)[keyof typeof lookups.authorizationStatuses];
    expires_at: objectTypes.timestampNumber;
    grant_type?: keyof typeof lookups.grantTypes;
    grant_type_value?: (typeof lookups.grantTypes)[keyof typeof lookups.grantTypes];
    scope: string;
    token_type: keyof typeof lookups.tokenTypes;
    token_type_value: (typeof lookups.tokenTypes)[keyof typeof lookups.tokenTypes];
    error?: keyof typeof lookups.authorizationErrors;
    error_value?: (typeof lookups.authorizationErrors)[keyof typeof lookups.authorizationErrors];
    error_description?: string;
    error_uri?: objectTypes.urlString;
    resourceURI: objectTypes.urlString;
    authorizationURI: objectTypes.urlString;
}
export interface BatchListContent extends BaseContent {
    contentType: 'BatchList';
    resources: objectTypes.urlString[];
}
export interface CustomerContent extends BaseContent {
    contentType: 'Customer';
    kind?: keyof typeof lookups.customerKinds;
    kind_value?: (typeof lookups.customerKinds)[keyof typeof lookups.customerKinds];
    specialNeed?: string;
    vip?: objectTypes.booleanString;
    pucNumber?: string;
    status?: objectTypes.GreenButtonStatus;
    priority?: {
        value?: string;
        dateTime?: objectTypes.timestampNumber;
        remark?: string;
        reason?: string;
    };
    locale?: string;
    customerName?: string;
}
export interface CustomerAccountContent extends BaseContent {
    contentType: 'CustomerAccount';
    type?: string;
    authorName?: string;
    createdDateTime?: objectTypes.timestampNumber;
    lastModifiedDateTime?: objectTypes.timestampNumber;
    revisionNumber?: string;
    electronicAddress?: objectTypes.GreenButtonElectronicAddress;
    subject?: string;
    title?: string;
    docStatus?: objectTypes.GreenButtonStatus;
    status?: objectTypes.GreenButtonStatus;
    comment?: string;
    billingCycle?: string;
    budgetBill?: string;
    lastBillAmount?: number;
    notifications?: objectTypes.GreenButtonCustomerAccountNotification[];
    contactInfo?: {
        streetAddress?: objectTypes.GreenButtonStreetAddress;
        postalAddress?: objectTypes.GreenButtonStreetAddress;
        phone1?: objectTypes.GreenButtonTelephoneNumber;
        phone2?: objectTypes.GreenButtonTelephoneNumber;
        electronicAddress?: objectTypes.GreenButtonElectronicAddress;
        organisationName?: string;
    };
    accountId?: string;
}
export interface CustomerAgreementContent extends BaseContent {
    contentType: 'CustomerAgreement';
    type?: string;
    authorName?: string;
    createdDateTime?: objectTypes.timestampNumber;
    lastModifiedDateTime?: objectTypes.timestampNumber;
    revisionNumber?: string;
    electronicAddress?: objectTypes.GreenButtonElectronicAddress;
    subject?: string;
    title?: string;
    docStatus?: objectTypes.GreenButtonStatus;
    status?: objectTypes.GreenButtonStatus;
    comment?: string;
    signDate?: objectTypes.timestampNumber;
    validityInterval?: objectTypes.GreenButtonDuration;
    loadMgmt?: string;
    isPrePay?: objectTypes.booleanString;
    shutOffDateTime?: objectTypes.timestampNumber;
    DemandResponseProgram?: objectTypes.GreenButtonDemandResponseProgram[];
    PricingStructures?: objectTypes.urlString[];
    currency?: keyof typeof lookups.currencies;
    currency_value?: (typeof lookups.currencies)[keyof typeof lookups.currencies];
    futureStatus?: objectTypes.GreenButtonStatus;
    agreementId?: string;
}
export interface ElectricPowerQualitySummaryContent extends BaseContent {
    summaryInterval: objectTypes.GreenButtonDuration;
    flickerPlt?: number;
    flickerPst?: number;
    harmonicVoltage?: number;
    longInterruptions?: number;
    mainsVoltage?: number;
    measurementProtocol?: number;
    powerFrequency?: number;
    rapidVoltageChanges?: number;
    shortInterruptions?: number;
    supplyVoltageDips?: number;
    supplyVoltageVariations?: number;
    tempOvervoltage?: number;
}
export interface IntervalBlockContent extends BaseContent {
    contentType: 'IntervalBlock';
    interval: objectTypes.GreenButtonDuration;
    IntervalReading: objectTypes.GreenButtonIntervalReading[];
}
export interface LocalTimeParametersContent extends BaseContent {
    contentType: 'LocalTimeParameters';
    dstEndRule: string;
    dstOffset: number;
    dstStartRule: string;
    tzOffset: number;
}
export interface MeterContent extends BaseContent {
    contentType: 'Meter';
    type?: string;
    utcNumber?: string;
    serialNumber?: string;
    lotNumber?: string;
    purchasePrice?: string;
    critical?: objectTypes.booleanString;
    electronicAddress?: objectTypes.GreenButtonElectronicAddress;
    lifecycle?: {
        manufacturedDate?: objectTypes.timestampNumber;
        purchaseDate?: objectTypes.timestampNumber;
        receivedDate?: objectTypes.timestampNumber;
        installationDate?: objectTypes.timestampNumber;
        removalDate?: objectTypes.timestampNumber;
        retiredDate?: objectTypes.timestampNumber;
    };
    acceptanceTest?: {
        type?: string;
        success?: objectTypes.booleanString;
        dateTime?: objectTypes.timestampNumber;
    };
    initialCondition?: string;
    initialLossOfLife?: number;
    isVirtual?: objectTypes.booleanString;
    isPan?: objectTypes.booleanString;
    installCode?: string;
    amrSystem?: string;
    formNumber?: string;
    MeterMultipliers?: objectTypes.GreenButtonMeterMultiplier[];
    intervalLength?: number;
}
export interface MeterReadingContent extends BaseContent {
    contentType: 'MeterReading';
}
export interface ReadingTypeContent extends BaseContent {
    contentType: 'ReadingType';
    accumulationBehaviour?: keyof typeof lookups.accumulationBehaviours;
    accumulationBehaviour_value?: (typeof lookups.accumulationBehaviours)[keyof typeof lookups.accumulationBehaviours];
    commodity?: keyof typeof lookups.commodities;
    commodity_value?: (typeof lookups.commodities)[keyof typeof lookups.commodities];
    consumptionTier?: number;
    currency?: keyof typeof lookups.currencies;
    currency_value?: (typeof lookups.currencies)[keyof typeof lookups.currencies];
    dataQualifier?: keyof typeof lookups.dataQualifiers;
    dataQualifier_value?: (typeof lookups.dataQualifiers)[keyof typeof lookups.dataQualifiers];
    defaultQuality?: keyof typeof lookups.readingQualities;
    defaultQuality_value?: (typeof lookups.readingQualities)[keyof typeof lookups.readingQualities];
    flowDirection?: keyof typeof lookups.flowDirections;
    flowDirection_value?: (typeof lookups.flowDirections)[keyof typeof lookups.flowDirections];
    intervalLength?: number;
    kind?: keyof typeof lookups.readingTypeKinds;
    kind_value?: (typeof lookups.readingTypeKinds)[keyof typeof lookups.readingTypeKinds];
    phase?: keyof typeof lookups.phaseCodes;
    phase_value?: (typeof lookups.phaseCodes)[keyof typeof lookups.phaseCodes];
    powerOfTenMultiplier?: keyof typeof lookups.powerOfTenMultipliers;
    powerOfTenMultiplier_value?: (typeof lookups.powerOfTenMultipliers)[keyof typeof lookups.powerOfTenMultipliers];
    timeAttribute?: keyof typeof lookups.timeAttributes;
    timeAttribute_value?: (typeof lookups.timeAttributes)[keyof typeof lookups.timeAttributes];
    tou?: number;
    uom?: keyof typeof lookups.unitsOfMeasurement;
    uom_value?: (typeof lookups.unitsOfMeasurement)[keyof typeof lookups.unitsOfMeasurement];
    cpp?: number;
    interharmonic?: number;
    measuringPeriod?: keyof typeof lookups.measuringPeriods;
    measuringPeriod_value?: (typeof lookups.measuringPeriods)[keyof typeof lookups.measuringPeriods];
    argument?: number;
}
// ...
export interface ServiceStatusContent extends BaseContent {
    contentType: 'ServiceStatus';
    currentStatus: keyof typeof lookups.currentStatuses;
    currentStatus_value?: (typeof lookups.currentStatuses)[keyof typeof lookups.currentStatuses];
}
// ...
export type GreenButtonContent = ApplicationInformationContent | AuthorizationContent | BatchListContent | CustomerContent | CustomerAccountContent | CustomerAgreementContent | ElectricPowerQualitySummaryContent | IntervalBlockContent | LocalTimeParametersContent | MeterContent | MeterReadingContent | ReadingTypeContent | ServiceLocationContent | ServiceStatusContent | ServiceSupplierContent | UsagePointContent | UsageSummaryContent;